requirements.txt
Flask==2.2.2
requests==2.28.2
python-dotenv==1.0.0
transformers==4.37.0
torch==2.1.0
